pyproject.toml
[tool.poetry]
name = "internet_consumption"
version = "0.1.0"
description = ""
authors = ["Unviray <unviray@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
flake8-blind-except = "^0.2.0"
flake8-bugbear = "^21.9.1"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.6.1"
flake8-debugger = "^4.0.0"
flake8-isort = "^4.0.0"
icecream = "^2.1.1"
pytest-django = "^4.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
